Move helper classes to utils directory
