+++ src/utils/eth-contract-call.spec.ts
@@ -2,7 +2,7 @@ import * as assert from "assert";
 // import axios from "axios"
 import * as moxios from "moxios";
 import {IContractCall} from "../interfaces";
-import {sandbox} from "../spec-helpers";
+import {sandbox} from "./spec-helpers";
 import {Waas} from "../waas";
 // import {Waas} from "../waas";
 import {callContractFunction} from "./eth-contract-call";
