+++ src/utils/polling-helper.spec.ts
